Context setup and updated app.js
